Replace Storybook example with own component
